iematch : if/in, maintain sort, merge to full data set
This commit implements if/in. Obs kept (match sample)will be matched and
then merged back to the full data set. Obs with missing in groudummy
will also be dropped from the match sample.

Extended missin values are added to the _match* vars to indicate why not
matched. For example, excluded by if/in or missing in groupdummy = .m,
control not matched = .c

Original sort is maintained and preserve/restore is implemented