iebaltab: added test regarding issue #109
If F-test for joint significance of balance variables is used, but neither the t-test or the normalized difference are displayed, then the columns for comparison between treatment arms are not created, and the F-test cannot be displayed.

This commit (1) displays an error message if option ftest is used along with nottest, and normdiff is not selected, (2) disables option ftest, (3) still creates the balance table.