iefolder : remove cap for commands to install
This segment list the commands needed for this project, and installs them if not installed. there was a `cap` on the row that installed the command not already installed. It is better if there is an error thrown if the command could not be installed, or if there was a type in the command name listed as required.

I discussed this with @eatorange