Add SETtest option
In light of journals like the AER disallowing the use of significance stars, one might want to display the difference column(s) as mean difference with its appropriate standard error to allow readers their own inference. Specifying the option SETtest does not work in combination with PTtest. I've implemented this and done a reasonable amount of testing, but of course the use of this code is as always at your own peril.